Criando e ajustando service para Produtos e adicionando logs

+++ src/main/java/com/zelly/encomendas/encomendaszelly/service/produto/dadosAtualizacaoProduto.java
@@ -1,4 +1,4 @@
 package com.zelly.encomendas.encomendaszelly.service.produto;
 
-public record dadosAtualizacaoProduto(Long id, String nome, tamanhoRoupa tamanho, Double valor, int quantidadeEstoque) {
+public record dadosAtualizacaoProduto(Long id, String nome, tamanhoRoupa tamanho, Double valor, Long quantidadeEstoque) {
 }
